Move DataTablePro from Atoms to Molecules
